create data module
