+++ AFQ/__init__.py
@@ -1,5 +1,4 @@
 from .version import version as __version__  # noqa
-from .data import *  # noqa
 from .utils import *  # noqa
 from .api import *  # noqa
