+++ AFQ/data/s3bids.py
@@ -21,126 +21,7 @@ from tqdm.auto import tqdm
 import nibabel as nib
 
 from bids import BIDSLayout
-from AFQ.data import afq_home, to_bids_description
-
-
-def fetch_hcp(subjects,
-              hcp_bucket='hcp-openaccess',
-              profile_name="hcp",
-              path=None,
-              study='HCP_1200',
-              aws_access_key_id=None,
-              aws_secret_access_key=None):
-    """
-    Fetch HCP diffusion data and arrange it in a manner that resembles the
-    BIDS [1]_ specification.
-
-    Parameters
-    ----------
-    subjects : list
-        Each item is an integer, identifying one of the HCP subjects
-    hcp_bucket : string, optional
-        The name of the HCP S3 bucket. Default: "hcp-openaccess"
-    profile_name : string, optional
-        The name of the AWS profile used for access. Default: "hcp"
-    path : string, optional
-        Path to save files into. Default: '~/AFQ_data'
-    study : string, optional
-        Which HCP study to grab. Default: 'HCP_1200'
-    aws_access_key_id : string, optional
-        AWS credentials to HCP AWS S3. Will only be used if `profile_name` is
-        set to False.
-    aws_secret_access_key : string, optional
-        AWS credentials to HCP AWS S3. Will only be used if `profile_name` is
-        set to False.
-
-    Returns
-    -------
-    dict with remote and local names of these files,
-    path to BIDS derivative dataset
-
-    Notes
-    -----
-    To use this function with its default setting, you need to have a
-    file '~/.aws/credentials', that includes a section:
-
-    [hcp]
-    AWS_ACCESS_KEY_ID=XXXXXXXXXXXXXXXX
-    AWS_SECRET_ACCESS_KEY=XXXXXXXXXXXXXXXX
-
-    The keys are credentials that you can get from HCP
-    (see https://wiki.humanconnectome.org/display/PublicData/How+To+Connect+to+Connectome+Data+via+AWS)  # noqa
-
-    Local filenames are changed to match our expected conventions.
-
-    .. [1] Gorgolewski et al. (2016). The brain imaging data structure,
-           a format for organizing and describing outputs of neuroimaging
-           experiments. Scientific Data, 3::160044. DOI: 10.1038/sdata.2016.44.
-    """
-    if profile_name:
-        boto3.setup_default_session(profile_name=profile_name)
-    elif aws_access_key_id is not None and aws_secret_access_key is not None:
-        boto3.setup_default_session(
-            aws_access_key_id=aws_access_key_id,
-            aws_secret_access_key=aws_secret_access_key)
-    else:
-        raise ValueError("Must provide either a `profile_name` or ",
-                         "both `aws_access_key_id` and ",
-                         "`aws_secret_access_key` as input to 'fetch_hcp'")
-
-    s3 = boto3.resource('s3')
-    bucket = s3.Bucket(hcp_bucket)
-
-    if path is None:
-        if not op.exists(afq_home):
-            os.mkdir(afq_home)
-        my_path = afq_home
-    else:
-        my_path = path
-
-    base_dir = op.join(my_path, study, 'derivatives', 'dmriprep')
-
-    if not os.path.exists(base_dir):
-        os.makedirs(base_dir, exist_ok=True)
-
-    data_files = {}
-    for subject in subjects:
-        # We make a single session folder per subject for this case, because
-        # AFQ api expects session structure:
-        sub_dir = op.join(base_dir, f'sub-{subject}')
-        sess_dir = op.join(sub_dir, "ses-01")
-        if not os.path.exists(sub_dir):
-            os.makedirs(os.path.join(sess_dir, 'dwi'), exist_ok=True)
-            os.makedirs(os.path.join(sess_dir, 'anat'), exist_ok=True)
-        data_files[op.join(sess_dir, 'dwi', f'sub-{subject}_dwi.bval')] =\
-            f'{study}/{subject}/T1w/Diffusion/bvals'
-        data_files[op.join(sess_dir, 'dwi', f'sub-{subject}_dwi.bvec')] =\
-            f'{study}/{subject}/T1w/Diffusion/bvecs'
-        data_files[op.join(sess_dir, 'dwi', f'sub-{subject}_dwi.nii.gz')] =\
-            f'{study}/{subject}/T1w/Diffusion/data.nii.gz'
-        data_files[op.join(sess_dir, 'anat', f'sub-{subject}_T1w.nii.gz')] =\
-            f'{study}/{subject}/T1w/T1w_acpc_dc.nii.gz'
-        data_files[op.join(sess_dir, 'anat',
-                           f'sub-{subject}_aparc+aseg_seg.nii.gz')] =\
-            f'{study}/{subject}/T1w/aparc+aseg.nii.gz'
-
-    for k in data_files.keys():
-        if not op.exists(k):
-            bucket.download_file(data_files[k], k)
-    # Create the BIDS dataset description file text
-    hcp_acknowledgements = """Data were provided by the Human Connectome Project, WU-Minn Consortium (Principal Investigators: David Van Essen and Kamil Ugurbil; 1U54MH091657) funded by the 16 NIH Institutes and Centers that support the NIH Blueprint for Neuroscience Research; and by the McDonnell Center for Systems Neuroscience at Washington University.""",  # noqa
-    to_bids_description(op.join(my_path, study),
-                        **{"Name": study,
-                           "Acknowledgements": hcp_acknowledgements,
-                           "Subjects": subjects})
-
-    # Create the BIDS derivatives description file text
-    to_bids_description(base_dir,
-                        **{"Name": study,
-                           "Acknowledgements": hcp_acknowledgements,
-                           "PipelineDescription": {'Name': 'dmriprep'}})
-
-    return data_files, op.join(my_path, study)
+from AFQ.data.fetch import afq_home, to_bids_description
 
 
 # +----------------------------------------------------+
