+++ AFQ/data/utils.py
@@ -1,0 +1,34 @@
+BUNDLE_RECO_2_AFQ = \
+    {
+        "AF_L": "ARC_L", "AF_R": "ARC_R",
+        "UF_L": "UNC_L", "UF_R": "UNC_R",
+        "IFOF_L": "IFO_L", "IFOF_R": "IFO_R",
+        "CST_L": "CST_L", "CST_R": "CST_R",
+        "ILF_L": "ILF_L", "ILF_R": "ILF_R",
+        "SLF_L": "SLF_L", "SLF_R": "SLF_R"
+    }
+
+BUNDLE_MAT_2_PYTHON = \
+    {'Right Corticospinal': 'CST_R', 'Left Corticospinal': 'CST_L',
+     'RightCorticospinal': 'CST_R', 'LeftCorticospinal': 'CST_L',
+     'Right Uncinate': 'UNC_R', 'Left Uncinate': 'UNC_L',
+     'RightUncinate': 'UNC_R', 'LeftUncinate': 'UNC_L',
+     'Left IFOF': 'IFO_L', 'Right IFOF': 'IFO_R',
+     'LeftIFOF': 'IFO_L', 'RightIFOF': 'IFO_R',
+     'Right Arcuate': 'ARC_R', 'Left Arcuate': 'ARC_L',
+     'RightArcuate': 'ARC_R', 'LeftArcuate': 'ARC_L',
+     'Right Thalamic Radiation': 'ATR_R', 'Left Thalamic Radiation': 'ATR_L',
+     'RightThalamicRadiation': 'ATR_R', 'LeftThalamicRadiation': 'ATR_L',
+     'Right Cingulum Cingulate': 'CGC_R', 'Left Cingulum Cingulate': 'CGC_L',
+     'RightCingulumCingulate': 'CGC_R', 'LeftCingulumCingulate': 'CGC_L',
+     'Right Cingulum Hippocampus': 'HCC_R',
+     'Left Cingulum Hippocampus': 'HCC_L',
+     'RightCingulumHippocampus': 'HCC_R',
+     'LeftCingulumHippocampus': 'HCC_L',
+     'Callosum Forceps Major': 'FP', 'Callosum Forceps Minor': 'FA',
+     'CallosumForcepsMajor': 'FP', 'CallosumForcepsMinor': 'FA',
+     'Right ILF': 'ILF_R', 'Left ILF': 'ILF_L',
+     'RightILF': 'ILF_R', 'LeftILF': 'ILF_L',
+     'Right SLF': 'SLF_R', 'Left SLF': 'SLF_L',
+     'RightSLF': 'SLF_R', 'LeftSLF': 'SLF_L'}
+
