+++ AFQ/tests/test_registration.py
@@ -12,7 +12,7 @@ from AFQ.registration import (register_dwi, write_mapping,
                               read_mapping, syn_register_dwi,
                               slr_registration)
 
-import AFQ.data as afd
+import AFQ.data.fetch as afd
 
 from dipy.io.streamline import load_tractogram
 from dipy.io.stateful_tractogram import Space
