+++ AFQ/utils/streamlines.py
@@ -2,7 +2,7 @@ from dipy.io.streamline import load_tractogram
 import numpy as np
 import nibabel as nib
 from dipy.io.stateful_tractogram import StatefulTractogram, Space
-from AFQ.s3bids import read_json
+from AFQ.data.s3bids import read_json
 import os.path as op
