+++ AFQ/utils/tests/test_volume.py
@@ -9,7 +9,7 @@ from dipy.io.stateful_tractogram import Space
 from dipy.io.streamline import StatefulTractogram
 
 import AFQ.utils.volume as afv
-import AFQ.data as afd
+import AFQ.data.fetch as afd
 
 
 def test_patch_up_roi():
