+++ examples/plot_afq_reco80.py
@@ -12,7 +12,7 @@ import os.path as op
 import plotly
 
 from AFQ.api.group import GroupAFQ
-import AFQ.data as afd
+import AFQ.data.fetch as afd
 
 ##########################################################################
 # Get some example data
