+++ examples/plot_optic_radiations.py
@@ -15,7 +15,7 @@ provide these ROIs as part of the software.
 import os.path as op
 from AFQ.api.group import GroupAFQ
 import AFQ.api.bundle_dict as abd
-import AFQ.data as afd
+import AFQ.data.fetch as afd
 from AFQ.definitions.mask import LabelledMaskFile, RoiMask
 
 afd.organize_stanford_data(clear_previous_afq=True)
